+++ examples/custom-esbuild/sanity-esbuild-app/src/app/app.component.ts
@@ -3,6 +3,7 @@ import { Component } from '@angular/core';
 declare const title: string;
 declare const subtitle: string;
 declare const titleByOption: string;
+declare const buildConfiguration: string;
 
 @Component({
   selector: 'app-root',
@@ -14,10 +15,12 @@ export class AppComponent {
   title: string;
   subtitle: string;
   titleByOption: string;
+  buildConfiguration: string;
 
   constructor() {
     this.title = typeof title !== 'undefined' ? title : 'sanity-esbuild-app';
     this.subtitle = typeof subtitle !== 'undefined' ? subtitle : 'sanity-esbuild-app subtitle';
     this.titleByOption = typeof titleByOption !== 'undefined' ? titleByOption : 'sanity-esbuild-app optionTitle';
+    this.buildConfiguration = typeof buildConfiguration !== 'undefined' ? buildConfiguration : 'no-config';
   }
 }