+++ packages/jest/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [21.0.4-beta.16](https://github.com/just-jeb/angular-builders/compare/@angular-builders/jest@21.0.4-beta.15...@angular-builders/jest@21.0.4-beta.16) (2026-06-05)
+
+### Reverts
+
+* remove redundant TS2742 builder annotations ([#2275](https://github.com/just-jeb/angular-builders/issues/2275), [#2278](https://github.com/just-jeb/angular-builders/issues/2278)) ([#2279](https://github.com/just-jeb/angular-builders/issues/2279)) ([a2882e5](https://github.com/just-jeb/angular-builders/commit/a2882e511ae2fa44dc445dbc9e73882de70981b5))
+
+## [21.0.4-beta.15](https://github.com/just-jeb/angular-builders/compare/@angular-builders/jest@21.0.4-beta.14...@angular-builders/jest@21.0.4-beta.15) (2026-06-04)
+
+### Bug Fixes
+
+* **builders:** annotate builder default exports with Builder<T> to avoid TS2742 ([#2278](https://github.com/just-jeb/angular-builders/issues/2278)) ([7db3848](https://github.com/just-jeb/angular-builders/commit/7db3848c7c3bf8362904130cbab8c7711cdac4ed))
+
+## [21.0.4-beta.14](https://github.com/just-jeb/angular-builders/compare/@angular-builders/jest@21.0.4-beta.13...@angular-builders/jest@21.0.4-beta.14) (2026-06-01)
+
+### Bug Fixes
+
+* **jest:** suppress warning when jest config is provided inline in angular.json (fixes [#1102](https://github.com/just-jeb/angular-builders/issues/1102)) ([#2213](https://github.com/just-jeb/angular-builders/issues/2213)) ([6e7c93f](https://github.com/just-jeb/angular-builders/commit/6e7c93f2e9a02ff4dafdbb0544fb0da7ace6afb2))
+
+## [21.0.4-beta.13](https://github.com/just-jeb/angular-builders/compare/@angular-builders/jest@21.0.4-beta.12...@angular-builders/jest@21.0.4-beta.13) (2026-06-01)
+
+### Bug Fixes
+
+* **jest:** emit --findRelatedTests with positional file args (fixes [#2150](https://github.com/just-jeb/angular-builders/issues/2150), [#1859](https://github.com/just-jeb/angular-builders/issues/1859)) ([#2237](https://github.com/just-jeb/angular-builders/issues/2237)) ([47f3b67](https://github.com/just-jeb/angular-builders/commit/47f3b67e2ee487f4c8d9019f1cd83dbf2e0c7284))
+
+## [21.0.4-beta.12](https://github.com/just-jeb/angular-builders/compare/@angular-builders/jest@21.0.4-beta.11...@angular-builders/jest@21.0.4-beta.12) (2026-06-01)
+
+**Note:** Version bump only for package @angular-builders/jest
+
 ## [21.0.4-beta.11](https://github.com/just-jeb/angular-builders/compare/@angular-builders/jest@21.0.4-beta.10...@angular-builders/jest@21.0.4-beta.11) (2026-05-09)
 
 **Note:** Version bump only for package @angular-builders/jest